@@ -18,10 +18,10 @@ class _AddDrinkDialogState extends State<AddDrinkDialog> {
   final _formKey = GlobalKey<FormState>();
   String name = '';
   String description = '';
+  String recipe = ''; // New field for recipe
   double price = 0.0;
-  XFile? imageFile; // Changed to XFile?
-  bool isAlcoholic = false; // New field to hold alcoholic status
-
+  XFile? imageFile; // Changed to XFile
+  bool isAlcoholic = false; // Field for alcoholic status
   bool isLoading = false;
 
   Future<void> pickImage() async {
@@ -34,6 +34,28 @@ class _AddDrinkDialogState extends State<AddDrinkDialog> {
         imageFile = pickedFile; // Now using XFile
       });
     }
+  }
+
+  // Function to generate search keywords from the drink name
+  List<String> generateSearchKeywords(String name) {
+    List<String> keywords = [];
+    String lowerCaseName = name.toLowerCase();
+
+    // Add all substrings of the name
+    for (int i = 1; i <= lowerCaseName.length; i++) {
+      keywords.add(lowerCaseName.substring(0, i));
+    }
+
+    // Split the name into words and add substrings of each word
+    List<String> words = lowerCaseName.split(' ');
+    for (String word in words) {
+      for (int i = 1; i <= word.length; i++) {
+        keywords.add(word.substring(0, i));
+      }
+    }
+
+    // Remove duplicates
+    return keywords.toSet().toList();
   }
 
   Future<void> addDrink() async {
@@ -66,19 +88,24 @@ class _AddDrinkDialogState extends State<AddDrinkDialog> {
       // Proceed with uploading the image and adding the drink
       String imageUrl = await uploadImage(imageFile!);
 
-      // Add the new drink to Firestore
-      String? userId = FirebaseAuth.instance.currentUser?.uid;
+      // Generate search keywords from the drink name
+      List<String> keywords = generateSearchKeywords(name);
 
-await FirebaseFirestore.instance.collection('drinks').add({
-  'averageRating': 0,
-  'description': description,
-  'imageUrl': imageUrl,
-  'name': name,
-  'price': price,
-  'isAlcoholic': isAlcoholic,
-  'reviews': [],
-  'createdBy': userId, // Add this line
-});
+      // Add the new drink to Firestore
+      String? userId = user.uid;
+
+      await FirebaseFirestore.instance.collection('drinks').add({
+        'averageRating': 0,
+        'description': description,
+        'imageUrl': imageUrl,
+        'name': name,
+        'recipe': recipe, // Save the recipe to Firestore
+        'price': price,
+        'isAlcoholic': isAlcoholic,
+        'reviews': [],
+        'searchKeywords': keywords, // Add this line
+        'createdBy': userId, // Add this line
+      });
 
       setState(() {
         isLoading = false;
@@ -107,34 +134,35 @@ await FirebaseFirestore.instance.collection('drinks').add({
   }
 
   Future<String> uploadImage(XFile imageFile) async {
-  try {
-    String fileName = DateTime.now().millisecondsSinceEpoch.toString();
-    Reference storageRef = FirebaseStorage.instance.ref().child('drink_images/$fileName.jpg');
+    try {
+      String fileName = DateTime.now().millisecondsSinceEpoch.toString();
+      Reference storageRef =
+          FirebaseStorage.instance.ref().child('drink_images/$fileName.jpg');
 
-    UploadTask uploadTask;
+      UploadTask uploadTask;
 
-    if (kIsWeb) {
-      Uint8List imageData = await imageFile.readAsBytes();
-      uploadTask = storageRef.putData(
-        imageData,
-        SettableMetadata(contentType: 'image/jpeg'),
-      );
-    } else {
-      File file = File(imageFile.path);
-      uploadTask = storageRef.putFile(
-        file,
-        SettableMetadata(contentType: 'image/jpeg'), // Include metadata
-      );
+      if (kIsWeb) {
+        Uint8List imageData = await imageFile.readAsBytes();
+        uploadTask = storageRef.putData(
+          imageData,
+          SettableMetadata(contentType: 'image/jpeg'),
+        );
+      } else {
+        File file = File(imageFile.path);
+        uploadTask = storageRef.putFile(
+          file,
+          SettableMetadata(contentType: 'image/jpeg'), // Include metadata
+        );
+      }
+
+      TaskSnapshot snapshot = await uploadTask;
+      String downloadUrl = await snapshot.ref.getDownloadURL();
+      return downloadUrl;
+    } catch (e) {
+      print('Error uploading image: $e');
+      throw Exception('Failed to upload image');
     }
-
-    TaskSnapshot snapshot = await uploadTask;
-    String downloadUrl = await snapshot.ref.getDownloadURL();
-    return downloadUrl;
-  } catch (e) {
-    print('Error uploading image: $e');
-    throw Exception('Failed to upload image');
   }
-}
 
   @override
   Widget build(BuildContext context) {
@@ -229,6 +257,25 @@ await FirebaseFirestore.instance.collection('drinks').add({
                     },
                   ),
                   const SizedBox(height: 15),
+                  // Recipe Field (New)
+                  TextFormField(
+                    maxLines: 3,
+                    decoration: const InputDecoration(
+                      labelText: 'Recipe',
+                      prefixIcon: Icon(Icons.receipt_long),
+                      border: OutlineInputBorder(),
+                    ),
+                    validator: (value) {
+                      if (value == null || value.isEmpty) {
+                        return 'Please enter the recipe';
+                      }
+                      return null;
+                    },
+                    onSaved: (value) {
+                      recipe = value!.trim();
+                    },
+                  ),
+                  const SizedBox(height: 15),
                   // Price Field
                   TextFormField(
                     decoration: const InputDecoration(
@@ -253,37 +300,37 @@ await FirebaseFirestore.instance.collection('drinks').add({
                   ),
                   const SizedBox(height: 15),
                   // Alcoholic Status Field
-                  // Using ToggleButtons for selection
-                  Row(
+                  // Using Wrap to ensure proper layout on small screens
+                  Wrap(
+                    crossAxisAlignment: WrapCrossAlignment.center,
+                    spacing: 10,
                     children: [
                       const Text(
                         'Alcohol Content:',
                         style: TextStyle(fontSize: 16),
                       ),
-                      const SizedBox(width: 10),
-                      Expanded(
-                        child: ToggleButtons(
-                          isSelected: [!isAlcoholic, isAlcoholic],
-                          borderRadius: BorderRadius.circular(10),
-                          selectedColor: Colors.white,
-                          fillColor: Theme.of(context).primaryColor,
-                          color: Colors.grey,
-                          children: const [
-                            Padding(
-                              padding: EdgeInsets.symmetric(horizontal: 16),
-                              child: Text('Non-Alcoholic'),
-                            ),
-                            Padding(
-                              padding: EdgeInsets.symmetric(horizontal: 16),
-                              child: Text('Alcoholic'),
-                            ),
-                          ],
-                          onPressed: (int index) {
-                            setState(() {
-                              isAlcoholic = index == 1;
-                            });
-                          },
-                        ),
+                      ToggleButtons(
+                        isSelected: [!isAlcoholic, isAlcoholic],
+                        borderRadius: BorderRadius.circular(10),
+                        selectedColor: Colors.white,
+                        fillColor: Theme.of(context).primaryColor,
+                        color: Colors.grey,
+                        constraints: const BoxConstraints(minHeight: 36),
+                        children: const [
+                          Padding(
+                            padding: EdgeInsets.symmetric(horizontal: 8),
+                            child: Text('Non-Alc'),
+                          ),
+                          Padding(
+                            padding: EdgeInsets.symmetric(horizontal: 8),
+                            child: Text('Alcoholic'),
+                          ),
+                        ],
+                        onPressed: (int index) {
+                          setState(() {
+                            isAlcoholic = index == 1;
+                          });
+                        },
                       ),
                     ],
                   ),
